更新easyui, 创建模型模板

+++ resources/assets/backend/static/js/easyui/src/jquery.form.js
@@ -1,56 +1,105 @@
 /**
- * form - jQuery EasyUI
+ * EasyUI for jQuery 1.7.0
  * 
- * Copyright (c) 2009-2013 www.jeasyui.com. All rights reserved.
+ * Copyright (c) 2009-2018 www.jeasyui.com. All rights reserved.
  *
- * Licensed under the GPL or commercial licenses
+ * Licensed under the freeware license: http://www.jeasyui.com/license_freeware.php
  * To use it on other terms please contact us: info@jeasyui.com
- * http://www.gnu.org/licenses/gpl.txt
- * http://www.jeasyui.com/license_commercial.php
+ *
+ */
+/**
+ * form - EasyUI for jQuery
+ * 
  */
 (function($){
 	/**
 	 * submit the form
 	 */
 	function ajaxSubmit(target, options){
-		options = options || {};
+		var opts = $.data(target, 'form').options;
+		$.extend(opts, options||{});
 		
-		var param = {};
-		if (options.onSubmit){
-			if (options.onSubmit.call(target, param) == false) {
-				return;
+		var param = $.extend({}, opts.queryParams);
+		if (opts.onSubmit.call(target, param) == false){return;}
+
+		// $(target).find('.textbox-text:focus').blur();
+		var input = $(target).find('.textbox-text:focus');
+		input.triggerHandler('blur');
+		input.focus();
+
+		var disabledFields = null;	// the fields to be disabled
+		if (opts.dirty){
+			var ff = [];	// all the dirty fields
+			$.map(opts.dirtyFields, function(f){
+				if ($(f).hasClass('textbox-f')){
+					$(f).next().find('.textbox-value').each(function(){
+						ff.push(this);
+					});
+				} else {
+					ff.push(f);
+				}
+			});
+			disabledFields = $(target).find('input[name]:enabled,textarea[name]:enabled,select[name]:enabled').filter(function(){
+				return $.inArray(this, ff) == -1;
+			});
+			// disabledFields.attr('disabled', 'disabled');
+			disabledFields._propAttr('disabled', true);
+		}
+
+		if (opts.ajax){
+			if (opts.iframe){
+				submitIframe(target, param);
+			} else {
+				if (window.FormData !== undefined){
+					submitXhr(target, param);
+				} else {
+					submitIframe(target, param);
+				}
 			}
+		} else {
+			$(target).submit();
 		}
-		
-		var form = $(target);
-		if (options.url){
-			form.attr('action', options.url);
+
+		if (opts.dirty){
+			// disabledFields.removeAttr('disabled');
+			disabledFields._propAttr('disabled', false);
 		}
+	}
+
+	function submitIframe(target, param){
+		var opts = $.data(target, 'form').options;
 		var frameId = 'easyui_frame_' + (new Date().getTime());
-		var frame = $('<iframe id='+frameId+' name='+frameId+'></iframe>')
-				.attr('src', window.ActiveXObject ? 'javascript:false' : 'about:blank')
-				.css({
-					position:'absolute',
-					top:-1000,
-					left:-1000
-				});
-		var t = form.attr('target'), a = form.attr('action');
-		form.attr('target', frameId);
+		var frame = $('<iframe id='+frameId+' name='+frameId+'></iframe>').appendTo('body')
+		frame.attr('src', window.ActiveXObject ? 'javascript:false' : 'about:blank');
+		frame.css({
+			position:'absolute',
+			top:-1000,
+			left:-1000
+		});
+		frame.bind('load', cb);
 		
-		var paramFields = $();
-		try {
-			frame.appendTo('body');
-			frame.bind('load', cb);
-			for(var n in param){
-				var f = $('<input type="hidden" name="' + n + '">').val(param[n]).appendTo(form);
-				paramFields = paramFields.add(f);
+		submit(param);
+		
+		function submit(param){
+			var form = $(target);
+			if (opts.url){
+				form.attr('action', opts.url);
 			}
-			checkState();
-			form[0].submit();
-		} finally {
-			form.attr('action', a);
-			t ? form.attr('target', t) : form.removeAttr('target');
-			paramFields.remove();
+			var t = form.attr('target'), a = form.attr('action');
+			form.attr('target', frameId);
+			var paramFields = $();
+			try {
+				for(var n in param){
+					var field = $('<input type="hidden" name="' + n + '">').val(param[n]).appendTo(form);
+					paramFields = paramFields.add(field);
+				}
+				checkState();
+				form[0].submit();
+			} finally {
+				form.attr('action', a);
+				t ? form.attr('target', t) : form.removeAttr('target');
+				paramFields.remove();
+			}
 		}
 		
 		function checkState(){
@@ -68,19 +117,18 @@
 		
 		var checkCount = 10;
 		function cb(){
-			var frame = $('#'+frameId);
-			if (!frame.length){return}
-			frame.unbind();
+			var f = $('#'+frameId);
+			if (!f.length){return}
+			f.unbind();
 			var data = '';
 			try{
-				var body = frame.contents().find('body');
+				var body = f.contents().find('body');
 				data = body.html();
 				if (data == ''){
 					if (--checkCount){
 						setTimeout(cb, 100);
 						return;
 					}
-//					return;
 				}
 				var ta = body.find('>textarea');
 				if (ta.length){
@@ -92,17 +140,49 @@
 					}
 				}
 			} catch(e){
-				
 			}
-			if (options.success){
-				options.success(data);
-			}
+			opts.success.call(target, data);
 			setTimeout(function(){
-				frame.unbind();
-				frame.remove();
+				f.unbind();
+				f.remove();
 			}, 100);
 		}
 	}
+
+	function submitXhr(target, param){
+		var opts = $.data(target, 'form').options;
+		var formData = new FormData($(target)[0]);
+		for(var name in param){
+			formData.append(name, param[name]);
+		}
+		$.ajax({
+			url: opts.url,
+			type: 'post',
+			xhr: function(){
+				var xhr = $.ajaxSettings.xhr();
+				if (xhr.upload) {
+					xhr.upload.addEventListener('progress', function(e){
+						if (e.lengthComputable) {
+							var total = e.total;
+							var position = e.loaded || e.position;
+							var percent = Math.ceil(position * 100 / total);
+							opts.onProgress.call(target, percent);
+						}
+					}, false);
+				}
+				return xhr;
+			},
+			data: formData,
+			dataType: 'html',
+			cache: false,
+			contentType: false,
+			processData: false,
+			complete: function(res){
+				opts.success.call(target, res.responseText);
+			}
+		});
+	}
+	
 	
 	/**
 	 * load form data
@@ -110,11 +190,6 @@
 	 * otherwise load from local data object. 
 	 */
 	function load(target, data){
-		if (!$.data(target, 'form')){
-			$.data(target, 'form', {
-				options: $.extend({}, $.fn.form.defaults)
-			});
-		}
 		var opts = $.data(target, 'form').options;
 		
 		if (typeof data == 'string'){
@@ -140,75 +215,73 @@
 			var form = $(target);
 			for(var name in data){
 				var val = data[name];
-				var rr = _checkField(name, val);
-				if (!rr.length){
-//					var f = form.find('input[numberboxName="'+name+'"]');
-//					if (f.length){
-//						f.numberbox('setValue', val);	// set numberbox value
-//					} else {
-//						$('input[name="'+name+'"]', form).val(val);
-//						$('textarea[name="'+name+'"]', form).val(val);
-//						$('select[name="'+name+'"]', form).val(val);
-//					}
-					var count = _loadOther(name, val);
-					if (!count){
-						$('input[name="'+name+'"]', form).val(val);
-						$('textarea[name="'+name+'"]', form).val(val);
-						$('select[name="'+name+'"]', form).val(val);
+				if (!_checkField(name, val)){
+					if (!_loadBox(name, val)){
+						form.find('input[name="'+name+'"]').val(val);
+						form.find('textarea[name="'+name+'"]').val(val);
+						form.find('select[name="'+name+'"]').val(val);
 					}
 				}
-				_loadCombo(name, val);
 			}
 			opts.onLoadSuccess.call(target, data);
-			validate(target);
+			form.form('validate');
 		}
 		
 		/**
 		 * check the checkbox and radio fields
 		 */
 		function _checkField(name, val){
-			var rr = $(target).find('input[name="'+name+'"][type=radio], input[name="'+name+'"][type=checkbox]');
-			rr._propAttr('checked', false);
-			rr.each(function(){
-				var f = $(this);
-				if (f.val() == String(val) || $.inArray(f.val(), $.isArray(val)?val:[val]) >= 0){
-					f._propAttr('checked', true);
-				}
-			});
-			return rr;
-		}
-		
-		function _loadOther(name, val){
-			var count = 0;
-			var pp = ['numberbox','slider'];
-			for(var i=0; i<pp.length; i++){
-				var p = pp[i];
-				var f = $(target).find('input['+p+'Name="'+name+'"]');
-				if (f.length){
-					f[p]('setValue', val);
-					count += f.length;
+			var plugins = ['switchbutton','radiobutton','checkbox'];
+			for(var i=0; i<plugins.length; i++){
+				var plugin = plugins[i];
+				var cc = $(target).find('['+plugin+'Name="'+name+'"]');
+				if (cc.length){
+					cc[plugin]('uncheck');
+					cc.each(function(){
+						if (_isChecked($(this)[plugin]('options').value, val)){
+							$(this)[plugin]('check');
+						}
+					});
+					return true;
 				}
 			}
-			return count;
+			var cc = $(target).find('input[name="'+name+'"][type=radio], input[name="'+name+'"][type=checkbox]');
+			if (cc.length){
+				cc._propAttr('checked', false);
+				cc.each(function(){
+					if (_isChecked($(this).val(), val)){
+						$(this)._propAttr('checked', true);
+					}
+				});
+				return true;
+			}
+			return false;
+		}
+		function _isChecked(v, val){
+			if (v == String(val) || $.inArray(v, $.isArray(val)?val:[val]) >= 0){
+				return true;
+			} else {
+				return false;
+			}
 		}
 		
-		function _loadCombo(name, val){
-			var form = $(target);
-			var cc = ['combobox','combotree','combogrid','datetimebox','datebox','combo'];
-			var c = form.find('[comboName="' + name + '"]');
-			if (c.length){
-				for(var i=0; i<cc.length; i++){
-					var type = cc[i];
-					if (c.hasClass(type+'-f')){
-						if (c[type]('options').multiple){
-							c[type]('setValues', val);
+		function _loadBox(name, val){
+			var field = $(target).find('[textboxName="'+name+'"],[sliderName="'+name+'"]');
+			if (field.length){
+				for(var i=0; i<opts.fieldTypes.length; i++){
+					var type = opts.fieldTypes[i];
+					var state = field.data(type);
+					if (state){
+						if (state.options.multiple || state.options.range){
+							field[type]('setValues', val);
 						} else {
-							c[type]('setValue', val);
+							field[type]('setValue', val);
 						}
-						return;
+						return true;
 					}
 				}
 			}
+			return false;
 		}
 	}
 	
@@ -217,13 +290,22 @@
 	 */
 	function clear(target){
 		$('input,select,textarea', target).each(function(){
+			if ($(this).hasClass('textbox-value')){return;}
 			var t = this.type, tag = this.tagName.toLowerCase();
 			if (t == 'text' || t == 'hidden' || t == 'password' || tag == 'textarea'){
 				this.value = '';
 			} else if (t == 'file'){
 				var file = $(this);
-				file.after(file.clone().val(''));
-				file.remove();
+				if (!file.hasClass('textbox-value')){
+					var newfile = file.clone().val('');
+					newfile.insertAfter(file);
+					if (file.data('validatebox')){
+						file.validatebox('destroy');
+						newfile.validatebox();
+					} else {
+						file.remove();
+					}
+				}
 			} else if (t == 'checkbox' || t == 'radio'){
 				this.checked = false;
 			} else if (tag == 'select'){
@@ -231,46 +313,33 @@
 			}
 			
 		});
-//		if ($.fn.combo) $('.combo-f', target).combo('clear');
-//		if ($.fn.combobox) $('.combobox-f', target).combobox('clear');
-//		if ($.fn.combotree) $('.combotree-f', target).combotree('clear');
-//		if ($.fn.combogrid) $('.combogrid-f', target).combogrid('clear');
 		
-		var t = $(target);
-		var plugins = ['combo','combobox','combotree','combogrid','slider'];
-		for(var i=0; i<plugins.length; i++){
-			var plugin = plugins[i];
-			var r = t.find('.'+plugin+'-f');
-			if (r.length && r[plugin]){
-				r[plugin]('clear');
+		var tmp = $();
+		var form = $(target);
+		var opts = $.data(target, 'form').options;
+		for(var i=0; i<opts.fieldTypes.length; i++){
+			var type = opts.fieldTypes[i];
+			var field = form.find('.'+type+'-f').not(tmp);
+			if (field.length && field[type]){
+				field[type]('clear');
+				tmp = tmp.add(field);
 			}
 		}
-		validate(target);
+		form.form('validate');
 	}
 	
 	function reset(target){
 		target.reset();
-		var t = $(target);
-//		if ($.fn.combo){t.find('.combo-f').combo('reset');}
-//		if ($.fn.combobox){t.find('.combobox-f').combobox('reset');}
-//		if ($.fn.combotree){t.find('.combotree-f').combotree('reset');}
-//		if ($.fn.combogrid){t.find('.combogrid-f').combogrid('reset');}
-//		if ($.fn.datebox){t.find('.datebox-f').datebox('reset');}
-//		if ($.fn.datetimebox){t.find('.datetimebox-f').datetimebox('reset');}
-//		if ($.fn.spinner){t.find('.spinner-f').spinner('reset');}
-//		if ($.fn.timespinner){t.find('.timespinner-f').timespinner('reset');}
-//		if ($.fn.numberbox){t.find('.numberbox-f').numberbox('reset');}
-//		if ($.fn.numberspinner){t.find('.numberspinner-f').numberspinner('reset');}
-		
-		var plugins = ['combo','combobox','combotree','combogrid','datebox','datetimebox','spinner','timespinner','numberbox','numberspinner','slider'];
-		for(var i=0; i<plugins.length; i++){
-			var plugin = plugins[i];
-			var r = t.find('.'+plugin+'-f');
-			if (r.length && r[plugin]){
-				r[plugin]('reset');
+		var form = $(target);
+		var opts = $.data(target, 'form').options;
+		for(var i=opts.fieldTypes.length-1; i>=0; i--){
+			var type = opts.fieldTypes[i];
+			var field = form.find('.'+type+'-f');
+			if (field.length && field[type]){
+				field[type]('reset');
 			}
 		}
-		validate(target);
+		form.form('validate');
 	}
 	
 	/**
@@ -278,28 +347,44 @@
 	 */
 	function setForm(target){
 		var options = $.data(target, 'form').options;
-		var form = $(target);
-		form.unbind('.form').bind('submit.form', function(){
-			setTimeout(function(){
-				ajaxSubmit(target, options);
-			}, 0);
-			return false;
+		$(target).unbind('.form');
+		if (options.ajax){
+			$(target).bind('submit.form', function(){
+				setTimeout(function(){
+					ajaxSubmit(target, options);
+				}, 0);
+				return false;
+			});
+		}
+		$(target).bind('_change.form', function(e, t){
+			if ($.inArray(t, options.dirtyFields) == -1){
+				options.dirtyFields.push(t);
+			}
+			options.onChange.call(this, t);
+		}).bind('change.form', function(e){
+			var t = e.target;
+			if (!$(t).hasClass('textbox-text')){
+				if ($.inArray(t, options.dirtyFields) == -1){
+					options.dirtyFields.push(t);
+				}
+				options.onChange.call(this, t);
+			}
 		});
+		setValidation(target, options.novalidate);
 	}
 	
-//	function validate(target){
-//		if ($.fn.validatebox){
-//			var box = $('.validatebox-text', target);
-//			if (box.length){
-//				box.validatebox('validate');
-////				box.trigger('focus');
-////				box.trigger('blur');
-//				var invalidbox = $('.validatebox-invalid:first', target).focus();
-//				return invalidbox.length == 0;
-//			}
-//		}
-//		return true;
-//	}
+	function initForm(target, options){
+		options = options || {};
+		var state = $.data(target, 'form');
+		if (state){
+			$.extend(state.options, options);
+		} else {
+			$.data(target, 'form', {
+				options: $.extend({}, $.fn.form.defaults, $.fn.form.parseOptions(target), options)
+			});
+		}
+	}
+	
 	function validate(target){
 		if ($.fn.validatebox){
 			var t = $(target);
@@ -312,29 +397,32 @@
 	}
 	
 	function setValidation(target, novalidate){
+		var opts = $.data(target, 'form').options;
+		opts.novalidate = novalidate;
 		$(target).find('.validatebox-text:not(:disabled)').validatebox(novalidate ? 'disableValidation' : 'enableValidation');
 	}
 	
 	$.fn.form = function(options, param){
 		if (typeof options == 'string'){
+			this.each(function(){
+				initForm(this);
+			});
 			return $.fn.form.methods[options](this, param);
 		}
 		
-		options = options || {};
 		return this.each(function(){
-			if (!$.data(this, 'form')){
-				$.data(this, 'form', {
-					options: $.extend({}, $.fn.form.defaults, options)
-				});
-			}
+			initForm(this, options);
 			setForm(this);
 		});
 	};
 	
 	$.fn.form.methods = {
+		options: function(jq){
+			return $.data(jq[0], 'form').options;
+		},
 		submit: function(jq, options){
 			return jq.each(function(){
-				ajaxSubmit(this, $.extend({}, $.fn.form.defaults, options||{}));
+				ajaxSubmit(this, options);
 			});
 		},
 		load: function(jq, data){
@@ -364,15 +452,45 @@
 			return jq.each(function(){
 				setValidation(this, false);
 			});
+		},
+		resetValidation: function(jq){
+			return jq.each(function(){
+				$(this).find('.validatebox-text:not(:disabled)').validatebox('resetValidation');
+			});
+		},
+		resetDirty: function(jq){
+			return jq.each(function(){
+				$(this).form('options').dirtyFields = [];
+			});
 		}
 	};
 	
+	$.fn.form.parseOptions = function(target){
+		var t = $(target);
+		return $.extend({}, $.parser.parseOptions(target, [
+			{ajax:'boolean',dirty:'boolean'}
+		]), {
+			url: (t.attr('action') ? t.attr('action') : undefined)
+		});
+	};
+	
 	$.fn.form.defaults = {
+		fieldTypes: ['tagbox','combobox','combotree','combogrid','combotreegrid','datetimebox','datebox','combo',
+		        'datetimespinner','timespinner','numberspinner','spinner',
+		        'slider','searchbox','numberbox','passwordbox','filebox','textbox','switchbutton','radiobutton','checkbox'],
+		novalidate: false,
+		ajax: true,
+		iframe: true,
+		dirty: false,
+		dirtyFields: [],
 		url: null,
+		queryParams: {},
 		onSubmit: function(param){return $(this).form('validate');},
+		onProgress: function(percent){},
 		success: function(data){},
 		onBeforeLoad: function(param){},
 		onLoadSuccess: function(data){},
-		onLoadError: function(){}
+		onLoadError: function(){},
+		onChange: function(target){}
 	};
 })(jQuery);
